[feat] #478 metronome_play 이벤트 발생

+++ Hanbae/Service/DTO/AnalyticsEvent.swift
@@ -8,7 +8,7 @@
 import Mixpanel
 
 enum AnalyticsEvent {
-    case metronomePlay(duration: Int, soundType: String)
+    case metronomePlay(duration: Double, soundType: String)
     
     var name: String {
         switch self {
